Add ResourceExtensions
